% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{check_keys}
\alias{check_keys}
\title{Check API keys}
\usage{
check_keys()
}
\value{
Nothing.
}
\description{
Check which API keys are set as environmental varibales, or in other words, which API calls you are set to use.
}
\note{
This does not check whether the API keys are valid, merely that they are set.
}
