% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnnid_category.R
\name{cnnid_category}
\alias{cnnid_category}
\title{Get resources of CNN Indonesia}
\usage{
cnnid_category(category = NULL)
}
\arguments{
\item{category}{Character string that contains the category for the API's
data base.}
}
\value{
A tibble
}
\description{
\code{cnnid_search} returns from CNN Indonesia based on category that we specify.
}
\examples{
\dontrun{
df <- cnn_category(category = "nasional")
}
}
