% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnnid_index.R
\name{cnnid_index}
\alias{cnnid_index}
\title{Get random resources of CNN Indonesia}
\usage{
cnnid_index()
}
\value{
A tibble
}
\description{
\code{cnnid_search} returns random news from CNN Indonesia.
}
\examples{
\dontrun{
df <- cnnid_index()
}
}
