% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnnid_search.R
\name{cnnid_search}
\alias{cnnid_search}
\title{Get resources of CNN Indonesia}
\usage{
cnnid_search(keyword)
}
\arguments{
\item{keyword}{Character string that contains the searchterm for the API's
data base.}
}
\value{
A tibble
}
\description{
\code{cnnid_search} returns from CNN Indonesia based on keyword that we specify.
}
\examples{
\dontrun{
df <- cnnid_search(keyword = "jokowi")
}
}
