% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appledaily.R
\name{get_appledaily}
\alias{get_appledaily}
\title{Get Appledaily content with urls}
\usage{
get_appledaily(urls, nb_core = "auto")
}
\arguments{
\item{urls}{Links of news.}

\item{nb_core}{Number of cores.}
}
\description{
Get Appledaily content with urls
}
\examples{
urls = c(
   "http://www.appledaily.com.tw/appledaily/article/finance/20160626/37284508/",
   "http://www.appledaily.com.tw/realtimenews/article/international/20160830/938362/"
 )

out <- get_appledaily(urls)
}
