% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinatimes..r
\name{getAllNewsChinaTimes}
\alias{getAllNewsChinaTimes}
\title{China Times}
\usage{
getAllNewsChinaTimes(dateWant, sleepBetween = 5, sleepTime = 1,
  printSts = TRUE)
}
\arguments{
\item{dateWant}{\code{"yyyymmdd"}.}
}
\value{
data.table object with columns:
}
\description{
China Times
}
\examples{
getAllNewsUdn("20161001")

}

