% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctee.R
\name{getAllNewsCT}
\alias{getAllNewsCT}
\title{CTEE 工商時報 (工商e報)}
\usage{
getAllNewsCT(dateWant, sleepBetween = 5, sleepTime = 1, printSts = TRUE)
}
\arguments{
\item{dateWant}{\code{"yyyymmdd"}.}
}
\value{
data.table object with columns:
}
\description{
CTEE 工商時報 (工商e報)
}
\examples{
getAllNewsUdn("20161001")

}
