% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edn.R
\name{getAllNewsEC}
\alias{getAllNewsEC}
\title{Economic Daily News 經濟日報}
\usage{
getAllNewsEC(sleepBetween = 5, sleepTime = 1, printSts = TRUE)
}
\arguments{
\item{dateWant}{\code{"yyyymmdd"}.}
}
\value{
data.table object with columns:
}
\description{
\link{http://edn.udn.com/}
}
\examples{
getAllNewsUdn("20161001")

}
