% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETtoday.r
\name{getAllNewsETtoday}
\alias{getAllNewsETtoday}
\title{ETtoday}
\usage{
getAllNewsETtoday(dateWant, sleepBetween = 5, sleepTime = 1,
  printSts = TRUE)
}
\arguments{
\item{dateWant}{\code{"yyyymmdd"}.}
}
\value{
data.table object with columns:
}
\description{
ETtoday
}

