% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_everything.R
\name{news_everything}
\alias{news_everything}
\title{Get resources of newsapi.org}
\usage{
news_everything(
  keyword = NULL,
  keywordInTitle = NULL,
  sources = NULL,
  domains = NULL,
  excludeDomains = NULL,
  from = NULL,
  to = NULL,
  language = NULL,
  sortBy = "publishedAt",
  page = 1,
  pageSize = 100,
  get_all = FALSE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{keyword}{Character string that contains the searchterm for the API's
data base. API supports advanced search parameters, see 'details'.
Either keyword or keywordInTitle must be specified.}

\item{keywordInTitle}{Character string that does the same as above \emph{within the
headline only}. API supports advanced search parameters, see 'details'.
Either keyword or keywordInTitle must be specified.}

\item{sources}{Character vector with with IDs of the news outlets
you want to focus on (e.g., c("usa-today", "spiegel-online")).}

\item{domains}{Character vector with domains that you want
to restrict your search to (e.g. c("bbc.com", "nytimes.com")).}

\item{excludeDomains}{Similar usage as with 'domains'. Will exclude these
domains from your search.}

\item{from}{Character string with start date of your search. Needs to conform
to one of the following lubridate order strings:
\code{"ymdHMs, ymdHMsz, ymd"}. See help for lubridate::parse_date_time.
If from is not specified, NewsAPI defaults to the oldest available date
(depends on your paid/unpaid plan from newsapi.org).}

\item{to}{Character string that marks the end date of your search. Needs to conform
to one of the following lubridate order strings:
\code{"ymdHMs, ymdHMsz, ymd"}. See help for lubridate::parse_date_time.
If \code{to} is not specified,
NewsAPI defaults to the most recent article available.}

\item{language}{Specifies the language of the articles of your search. Must
be in ISO shortcut format (e.g., "de", "en"). Default
is all languages.}

\item{sortBy}{Character string that specifies the sorting variable of your article
results. Accepts three options: "publishedAt", "relevancy",
"popularity". Default is "publishedAt".}

\item{page}{Specifies the page number of your results that is returned. Must
be numeric. Default is first page.}

\item{pageSize}{The number of articles per page that are returned.
Maximum is 100 (also default).}

\item{get_all}{Logical value. If all pages will be collected then gel_all is TRUE}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, function can be
provided from the global environment (see \code{set_api_key()}).}
}
\value{
A tibble
}
\description{
\code{news_everything} returns articles from large and small news
sources and blogs. This includes news as well as other regular articles.
You can search for multiple \code{sources}, different \code{language},
or use your own keywords. Articles can be sorted by the earliest date
\code{publishedAt}, \code{relevancy}, or \code{popularity}. \cr\cr
Valid languages for \code{language} are provided in the dataset
\code{terms_language}.
}
\details{
Advanced search (see also www.newsapi.org): Surround entire phrases
with quotes (") for exact matches. Prepend words/phrases that must
appear with "+" symbol (e.g., +bitcoin). Prepend words that must not
appear with "-" symbol (e.g., -bitcoin). You can also use AND, OR,
NOT keywords (optionally grouped with parenthesis, e.g., 'crypto AND
(ethereum OR litecoin) NOT bitcoin)'). \cr
}
\examples{
\dontrun{
df <- news_everything(keyword = "jokowi", language = "id")
df <- news_everything(keyword = "komodo", from = "2019-01-02 12:00:00", get_all = TRUE)
}
}
