% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_headlines.R
\name{news_headlines}
\alias{news_headlines}
\title{Returns news headlines from newsapi.org}
\usage{
news_headlines(
  keyword = NULL,
  category = NULL,
  country = NULL,
  sources = NULL,
  page = 1,
  pageSize = 100,
  get_all = FALSE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{keyword}{Character string that contains the searchterm.}

\item{category}{Character string with the category you want headlines from.}

\item{country}{Character string with the country you want headlines from.}

\item{sources}{Character vector with with IDs of the news outlets
you want to focus on (e.g., c("usa-today", "spiegel-online")).}

\item{page}{Specifies the page number of your results that is returned. Must
be numeric. Default is first page. If you want to get all results
at once, use \code{get_headlines_all} from 'newsanchor'.}

\item{pageSize}{The number of articles per page that are returned.
Maximum is 100 (also default).}

\item{get_all}{Logical value. If all pages will be collected then gel_all is TRUE}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{news_headlines} returns headlines news for a country,
specific category in a country, single source, or multiple sources. You can
also search with keyword.
}
\examples{
\dontrun{
df <- news_headlines(sources = "bbc-news")
df <- news_headlines(keyword = "sports", page = 2)
df <- news_headlines(category = "business")
df <- news_headlines(keyword = "jokowi", country = "id", get_all = TRUE)
}
}
