% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_sources.R
\name{news_sources}
\alias{news_sources}
\title{Returns sources from newsapi.org}
\usage{
news_sources(api_key = Sys.getenv("NEWS_API_KEY"))
}
\arguments{
\item{api_key}{String with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{news_sources} returns the news sources currently available on newsapi.org.
}
\examples{
\dontrun{
news_sources(api_key)
}

}
