% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsapiContent.R
\name{newsapiContent}
\alias{newsapiContent}
\title{newsapiArticles}
\usage{
newsapiContent(url, tag)
}
\arguments{
\item{url}{(required) the url of the page from which to get the content}

\item{tag}{(required) the html tag from which to get the content
e.g. "p" passed to \code{\link[httr]{GET}}.}
}
\description{
A function to retrieve articles from newsapi.org
}
\keyword{api}
\keyword{news,}
