% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsapiSources.R
\name{newsapiSources}
\alias{newsapiSources}
\title{newsapiSources}
\usage{
newsapiSources(language = "", country = "", category = "")
}
\arguments{
\item{language}{(optional) - The 2-letter ISO-639-1 code of the language you would like to get sources for.
Possible options: "en", "de", "fr." Default: empty (all sources returned)}

\item{country}{(optional) - The 2-letter ISO 3166-1 code of the country you would like to get sources for.
Possible options: "au", "de", "gb", "in", "it", "us". Default: empty (all sources returned)}

\item{category}{(optional) - The category you would like to get sources for.
Possible options: business, entertainment, gaming, general, music, politics, science-and-nature, sport, technology.
Default: empty (all sources returned)}
}
\description{
newsapiSources
}
\keyword{api}
\keyword{news,}
