% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_aggregation.R
\name{newscatcher_aggregation}
\alias{newscatcher_aggregation}
\title{Returns document count from specific keywords}
\usage{
newscatcher_aggregation(
  keyword = NULL,
  topic = NULL,
  ranked_only = NULL,
  sources = NULL,
  agg_by = NULL,
  country = NULL,
  from = NULL,
  to_rank = NULL,
  lang = NULL,
  not_sources = NULL,
  to = NULL,
  from_rank = NULL,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{keyword}{Character string that contains the searchterm.}

\item{topic}{The topic to which you want to restrict the articles of your choice. This parameter is experimental.}

\item{ranked_only}{Logical value to limit the search only for the sources which are in top 1 million online websites. Unranked sources are assigned a rank that equals to 999999.}

\item{sources}{Character vector with one or more news resources to filter your search. It should be the normal form of the URL, for example, nytimes.com, theguardian.com.}

\item{agg_by}{How you want to aggregate the articles that math your query. Possible value: day or hour.}

\item{country}{Country to which you want to narrow your search. This parameter is experimental. We advise you to use it in conjunction with the lang parameter.}

\item{from}{Character that specify point in time to start the search, E.g. ("2020/05/01", "2020-05-01" ,"2020-05-01 12:54:14", "2020-05-01 4:15am", "yesterday 2:02 am", "2020/05/01 12:55 EST").}

\item{to_rank}{Upper boundary of the rank of news website to filter by. A number between 0 and 999999.}

\item{lang}{Specifies the language of the search.}

\item{not_sources}{One or more sources to be excluded from the search.}

\item{to}{Until which point in time to search for.}

\item{from_rank}{Lowest boundary of the rank of news website to filter by. Important: lower rank means that a source is more popular.}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{newscatcher_aggregation} returns document aggregation form specific keyword and filter.
Please note for free api you only get 5 articles per call. Free api limit the request per hour up to  21 request/hour
}
\examples{
\dontrun{df <- newscatcher_aggregation(keyword = "BNI", from = "2020-10-26")}

}
