% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_headlines.R
\name{newscatcher_headlines}
\alias{newscatcher_headlines}
\title{Returns lastest headlinse from specific filter}
\usage{
newscatcher_headlines(
  topic = NULL,
  lang = NULL,
  country = NULL,
  media = TRUE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{topic}{The topic to which you want to restrict the articles of your choice. This parameter is experimental.}

\item{lang}{Specifies the language of the search.}

\item{country}{Country to which you want to narrow your search. This parameter is experimental. We advise you to use it in conjunction with the lang parameter.}

\item{media}{Adds to the output of the call two more variables: media and media_content.\preformatted{        * media - the main image published with an article
        * media_content - a comma-separated string of all images used in an article
}}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{newscatcher_headlines} returns document of latest headlines form specific filter.
}
\examples{
\dontrun{df <- newscatcher_headlines(topic = "news", country = "id")}

}
