% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_search.R
\name{newscatcher_search}
\alias{newscatcher_search}
\title{Returns document from specific keywords}
\usage{
newscatcher_search(
  keyword = NULL,
  ranked_only = NULL,
  search_in = NULL,
  not_sources = NULL,
  sort_by = "relevancy",
  sources = NULL,
  to_rank = NULL,
  to = NULL,
  from = NULL,
  topic = NULL,
  page_size = 100,
  lang = NULL,
  country = NULL,
  page = 1,
  from_rank = NULL,
  get_all = FALSE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{keyword}{Character string that contains the searchterm.}

\item{ranked_only}{Logical value to limit the search only for the sources which are in top 1 million online websites. Unranked sources are assigned a rank that equals to 999999.}

\item{search_in}{By default, the function search what you specified in keyword argument in both title and summary of the article. However, you can limit this to either title or summary}

\item{not_sources}{One or more sources to be excluded from the search.}

\item{sort_by}{A character that specify sort method. Possible value:\preformatted{           * "relevancy" — the most relevant results first,
           * "date" — the most recently published results first,
           * "rank" — the results from the highest-ranked sources first
}}

\item{sources}{Character vector with one or more news resources to filter your search. It should be the normal form of the URL, for example, nytimes.com, theguardian.com.}

\item{to_rank}{Upper boundary of the rank of news website to filter by. A number between 0 and 999999.}

\item{to}{Until which point in time to search for.}

\item{from}{Character that specify point in time to start the search, E.g. ("2020/05/01", "2020-05-01" ,"2020-05-01 12:54:14", "2020-05-01 4:15am", "yesterday 2:02 am", "2020/05/01 12:55 EST").}

\item{topic}{The topic to which you want to restrict the articles of your choice. This parameter is experimental.}

\item{page_size}{How many articles to return per page. Defaults to 50, max is 100.}

\item{lang}{Logical value. Specifies the language of the search.}

\item{country}{Country to which you want to narrow your search. This parameter is experimental. We advise you to use it in conjunction with the lang parameter.}

\item{page}{The number of the page. Use it to scroll through the results. Defaults to 1.}

\item{from_rank}{Lowest boundary of the rank of news website to filter by. Important: lower rank means that a source is more popular.}

\item{get_all}{Logical value. If all pages will be collected then gel_all is TRUE.}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{newscatcher_search} returns document form specific keyword and filter.
Free api can only return 5 document, so make sure to specify page_size = 5.
Another things to remember is free api can only make 21 request/hour.
If you want to retrive another document more than 5 document please write a loop that only return 5 document/ request and sleep at least one hour after 21 request.
}
\examples{
\dontrun{
df <- newscatcher_search(keyword = "BNI", page_size = 5)}

}
