% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_searchfree.R
\name{newscatcher_searchfree}
\alias{newscatcher_searchfree}
\title{Returns document from specific keywords}
\usage{
newscatcher_searchfree(
  keyword = NULL,
  ranked_only = NULL,
  page_size = 50,
  media = TRUE,
  lang = NULL,
  page = 1,
  get_all = FALSE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{keyword}{Character string that contains the searchterm.}

\item{ranked_only}{Logical value to limit the search only for the sources which are in top 1 million online websites. Unranked sources are assigned a rank that equals to 999999.}

\item{page_size}{How many articles to return per page. Defaults to 50, max is 100.}

\item{media}{Adds to the output of the call two more variables: media and media_content.\preformatted{        * media - the main image published with an article
        * media_content - a comma-separated string of all images used in an article
}}

\item{lang}{Logical value. Specifies the language of the search.}

\item{page}{The number of the page. Use it to scroll through the results. Defaults to 1.}

\item{get_all}{Logical value. If all pages will be collected then gel_all is TRUE.}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{newscatcher_searchfree} returns document form specific keyword and filter.
The function can return up to 100 document per call despite using free api.
}
\examples{
\dontrun{df <- newscatcher_aggregation(keyword = "BNI", from = "2020-10-26")}

}
