% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_sources.R
\name{newscatcher_sources}
\alias{newscatcher_sources}
\title{Returns news sources from newscatcher api}
\usage{
newscatcher_sources(
  topic = NULL,
  lang = NULL,
  country = NULL,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{topic}{The topic to which you want to restrict the articles of your choice. This parameter is experimental. Accepted values are news, sport, tech, world, finance, politics, business, economics, entertainment.}

\item{lang}{Specifies the language of the search. Allowed values are: af, ar, bg, bn, ca,cn, cs, cy, da, de, el, en, es, et, fa, fi, fr, gu, he, hi, hr, hu, id, it, ja, kn, ko, lt, lv, mk, ml, mr, ne, nl, no, pa, pl, pt, ro, ru, sk, sl, so, sq, sv, sw, ta, te, th, tl, tr,tw, uk, ur, vi. Specifying the language will make your search more relevant}

\item{country}{The country to which you want to narrow your search. This parameter is experimental. We advise you to use it in conjunction with the lang parameter. Accepts any https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2.}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A character vector
}
\description{
\code{newscatcher_sources} returns news sources from newscatcher api.
}
\examples{
\dontrun{df <- newscatcher_sources(topic = "news", lang = "id")}

}
