% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcher_stock.R
\name{newscatcher_stocks}
\alias{newscatcher_stocks}
\title{Returns document from specific ticker}
\usage{
newscatcher_stocks(
  ticker = NULL,
  key_people = NULL,
  not_sources = NULL,
  from = NULL,
  to_rank = NULL,
  page_size = 100,
  from_rank = NULL,
  stock = NULL,
  to = NULL,
  sort_by = "relevancy",
  sources = NULL,
  page = 1,
  media = NULL,
  lang = NULL,
  get_all = FALSE,
  api_key = Sys.getenv("NEWS_API_KEY")
)
}
\arguments{
\item{ticker}{Ticker of a company you want to make your search about. Pure ticker without specifying its Stock Exchange. eg: AAPL for Apple Inc. ticker.}

\item{key_people}{Logical value to to match an article when it contains a name of one of the company C-level officer.}

\item{not_sources}{One or more sources to be excluded from the search.}

\item{from}{Character that specify point in time to start the search, E.g. ("2020/05/01", "2020-05-01" ,"2020-05-01 12:54:14", "2020-05-01 4:15am", "yesterday 2:02 am", "2020/05/01 12:55 EST").}

\item{to_rank}{Upper boundary of the rank of news website to filter by. A number between 0 and 999999.}

\item{page_size}{How many articles to return per page. Defaults to 50, max is 100.}

\item{from_rank}{Lowest boundary of the rank of news website to filter by. Important: lower rank means that a source is more popular.}

\item{stock}{Specify the Stock Exchange of the ticker that you look for. We support tickers from this list:\preformatted{          * ASX
          * AMEX
          * NYSE
          * NASDAQ
          * EURONEXT
          * HKEX
          * LSE
          * SGX
          * SIX
          * TSX
          * TSXV

          We really encourage you precise your Stock Exchange, because there are identical tickers among them.
}}

\item{to}{Until which point in time to search for.}

\item{sort_by}{A character that specify sort method. Possible value:\preformatted{           * "relevancy" — the most relevant results first,
           * "date" — the most recently published results first,
           * "rank" — the results from the highest-ranked sources first
}}

\item{sources}{Character vector with one or more news resources to filter your search. It should be the normal form of the URL, for example, nytimes.com, theguardian.com.}

\item{page}{The number of the page. Use it to scroll through the results. Defaults to 1.}

\item{media}{Adds to the output of the call two more variables: media and media_content.\preformatted{        * media - the main image published with an article
        * media_content - a comma-separated string of all images used in an article
}}

\item{lang}{Logical value. Specifies the language of the search.}

\item{get_all}{Logical value. If all pages will be collected then gel_all is TRUE.}

\item{api_key}{Character string with the API key you get from newsapi.org.
Passing it is compulsory. Alternatively, a function can be
provided from the global environment.}
}
\value{
A tibble
}
\description{
\code{newscatcher_stocks} returns document form specific ticker and stocks
Free api can only return 5 document, so make sure to specify page_size = 5.
Another things to remember is free api can only make 21 request/hour.
If you want to retrive another document more than 5 document please write a loop that only return 5 document/ request and sleep at least one hour after 21 request.
}
\examples{
\dontrun{
df <- newscatcher_stocks(ticker = "AAPL", page_size = 5)}

}
