% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsapiContent.R
\name{newsContent}
\alias{newsContent}
\title{newsContent}
\usage{
newsContent(url, tag, removeStop = "FALSE")
}
\arguments{
\item{url}{(required) the url of the page from which to get the content}

\item{tag}{(required) the html tag from which to get the content
e.g. "p" passed to \code{\link[xml2]{read_html}}.}

\item{removeStop}{(logical) remove stopwords from the raw text with an antijoin to
tidytext::stopwords? defaults to "FALSE".}
}
\description{
A function to retrieve articles from newsapi.org
}
\keyword{api}
\keyword{news,}
