% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsr.R
\docType{package}
\name{newsr}
\alias{newsr}
\alias{newsr-package}
\title{newsr: a Taiwan news crawler.}
\description{
newsr provides an API for accessing news websites.
}
\details{
It has following news resources:

\itemize{
\item Apple Daily 蘋果日報
\item Liberty Times Net 自由時報
\item UDN 聯合新聞網
\item China Times 中時電子報
\item ETtoday 東森新聞雲
}
}
\author{
LU YI
}

