% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appledaily.R
\name{search_appledaily}
\alias{search_appledaily}
\title{Search appledaily}
\usage{
search_appledaily(keyword, date_from = "2003-05-02", date_to = Sys.Date())
}
\arguments{
\item{keyword}{Vector of search strings.}

\item{date_from}{Begin of date of news to get with format \code{"yyyy-mm-dd"}.}

\item{date_to}{End of date of news to get with format \code{"yyyy-mm-dd"}.}
}
\value{
data.table object with columns:
}
\description{
Search appledaily
}
\examples{
search_appledaily(c("老年", "保險"))

}

