% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udn.r
\name{getAllNewsUdn}
\alias{getAllNewsUdn}
\title{UDN}
\usage{
getAllNewsUdn(dateWant, sleepBetween = 5, sleepTime = 1, printSts = TRUE)
}
\arguments{
\item{keyword}{Vector of search strings.}

\item{date_from}{Begin of date of news to get with format \code{"yyyy-mm-dd"}.}

\item{date_to}{End of date of news to get with format \code{"yyyy-mm-dd"}.}
}
\value{
data.table object with columns:
}
\description{
UDN
}
\examples{
getAllNewsUdn("20161001")

}

